requirements.txt
fastapi
uvicorn
httpx
gspread
google-auth-oauthlib
json